# openapi_agent_workflow_platform_mvp_v1.yaml
openapi: 3.0.3
info:
  title: Agent Workflow Platform MVP API
  version: 0.1.0
  description: MVP API draft for workflow authoring, publishing, running, tracing, knowledge bases, tools, models, and secrets.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /secrets:
    post:
      tags: [Secrets]
      summary: Create secret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSecretRequest"
      responses:
        "200":
          description: Secret metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
        default:
          $ref: "#/components/responses/ErrorResponse"
    get:
      tags: [Secrets]
      summary: List secrets without values
      responses:
        "200":
          description: Secret list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Secret"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /secrets/{secret_id}:
    put:
      tags: [Secrets]
      summary: Update secret
      parameters:
        - name: secret_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSecretRequest"
      responses:
        "200":
          description: Secret metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /node-types:
    get:
      tags: [Node Types]
      summary: List node types
      responses:
        "200":
          description: Node type list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/NodeTypeSummary"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /node-types/{node_type}/schema:
    get:
      tags: [Node Types]
      summary: Get node type schema
      parameters:
        - name: node_type
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/NodeType"
      responses:
        "200":
          description: Node schema
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeTypeSchema"
        default:
          $ref: "#/components/responses/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
    Timestamp:
      type: string
      format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              $ref: "#/components/schemas/JsonObject"
        request_id:
          type: string
...
    NodeType:
      type: string
      enum: [start, input, llm, knowledge_base, intent, branch, api, message, output, end]
...
    Secret:
      type: object
      properties:
        id:
          type: integer
          format: int64
        secret_key:
          type: string
        display_name:
          type: string
        status:
          type: string
        created_at:
          $ref: "#/components/schemas/Timestamp"
    CreateSecretRequest:
      type: object
      required: [secret_key, value]
      properties:
        secret_key:
          type: string
        display_name:
          type: string
        value:
          type: string
          format: password
    UpdateSecretRequest:
      type: object
      properties:
        display_name:
          type: string
        value:
          type: string
          format: password
    NodeTypeSummary:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/NodeType"
        name:
          type: string
        category:
          type: string
        description:
          type: string
    NodeTypeSchema:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/NodeType"
        node_schema:
          $ref: "#/components/schemas/JsonObject"
        config_schema:
          $ref: "#/components/schemas/JsonObject"
        form_schema:
          $ref: "#/components/schemas/JsonObject"